+++ filament/src/details/ShadowMapManager.h
@@ -86,6 +86,7 @@ private:
     struct ShadowLayout {
         uint8_t layer = 0;
         uint32_t size = 0;
+        uint8_t vsmSamples = 1;
     };
 
     struct TextureRequirements {