+++ RELEASE_NOTES.md
@@ -22,6 +22,8 @@ A new header is inserted each time a *tag* is created.
 - Add float versions of math constants to libmath
 - filament-utils: fix, `CoroutineScope` job should be canceled before destroy
 
+- New `ShadowOptions` control to render Variance Shadow Maps (VSM) with MSAA (experimental).
+
 ## v1.9.0
 
 - `MASKED` mode now leaves destination alpha intact (useful for transparent targets).